requirements/base.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
argh==0.26.2
attrs==16.0.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
billiard==3.3.0.23
celery==3.1.23
cffi==1.7.0
click==6.6
codecov==2.0.5
colorama==0.3.7
coverage==4.1
cryptography==1.4
cssselect==0.9.2
decorator==4.0.10
dj-database-url==0.4.1
Django==1.9.7
django-appconf==1.0.2
django-configurations==1.0
django-crispy-forms==1.6.0
django-extensions==1.6.7
django-filter==0.13.0
django-fsm==2.4.0
django-fsm-admin==1.2.3
django-fsm-log==1.3.0
djangorestframework==3.3.3
docopt==0.6.2
elasticsearch==1.9.0
enum34==1.1.6
envdir==0.7
fake-useragent==0.0.8
feedfinder2==0.0.4
feedparser==5.2.1
first==2.0.1
idna==2.1
ipaddress==1.0.16
ipdb==0.10.1
ipython==5.0.0
ipython-genutils==0.1.0
jieba3k==0.35.1
kombu==3.0.35
lxml==3.6.0
mccabe==0.5.0
newspaper3k==0.1.7
nltk==3.2.1
parsel==1.0.2
pathlib2==2.1.0
pathtools==0.1.2
pep257==0.7.0
pep8==1.7.0
pexpect==4.2.0
pickleshare==0.7.2
Pillow==3.3.0
pip-tools==1.7.0
prompt-toolkit==1.0.3
psycopg2==2.6.2
ptyprocess==0.5.1
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
pyflakes==1.2.3
Pygments==2.1.3
pylama==7.0.9
pyOpenSSL==16.0.0
pytest==2.9.2
pytest-capturelog==0.7
pytest-cov==2.3.0
pytest-django==2.9.1
pytest-testmon==0.8.2
pytest-watch==4.1.0
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
queuelib==1.4.2
requests==2.10.0
requests-file==1.4
Scrapy==1.1.0
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
tldextract==2.0.1
traitlets==4.2.2
Twisted==16.3.0
w3lib==1.14.2
watchdog==0.8.3
wcwidth==0.1.7
zope.interface==4.2.0
pika==0.10.0
